bot/requirements.txt
discord.py[voice]
websockets
youtube-dl
psycopg2
sqlalchemy
aiopg